getimagesize() failure
If getimagesize() fails, couse a user uploaded a none image file, a php
E_NOTICE is thrown.
I added validateImageSize() to output a custom error message, isntead of
throwing a E_NOTICE
It is used in upload().
If it returns false, the error message "mimeTypes" will be set and
upload() returns null.